% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablaFrecuencias.R
\name{num}
\alias{num}
\title{Tabla de Frecuencias}
\usage{
num(data)
}
\arguments{
\item{data}{base de datos}
}
\value{
tabla de frecuencias

Numero de datos totales "n"
Total number of data "n"
}
\description{
El siguiente codigo permite realizar una tabla de frecuencias basado en los metodos
estudiados en estadistica basica y disenio experimental

The next code allows create a frecuency table baase in the metodology studied at
DoE and stadistics
}
\author{
Michelle Paredes Escobar
}
