% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablaFrecuencias.R
\name{numStorge}
\alias{numStorge}
\title{Numero de categorias por el metodo de Storges
Number of categories. Storges}
\usage{
numStorge(num)
}
\description{
Numero de categorias por el metodo de Storges
Number of categories. Storges
}
